[BUGFIX] Language and renderType (#319)
* Added missing "renderType"

Adding the renderType for tax_class_id. Without this, the error occurs in the backend under 10.4.

* Correction of missing labels

The missing labels in the backend were corrected and adapted to the current state.

* Correction of missing labels

The missing labels in the backend were corrected and adapted to the current state.

* Default value of coupon_type field adjusted

Default value of the coupon_type field adjusted to the existing entry 'cartdiscount'.