[BUGFIX] Replace setTemplate Method with own View and setStep Method
The setTemplate() method not only sets the file to be rendered,
but overrides the action in the rendering context. This adjusted
the referrer in the form to an action that does not exist.
In the case of validation error, it is then incorrectly resolved
and an exception is thrown.
Through the own ViewTemplate it is possible to change only the
file name of the template to be rendered and the referrer is set
to the correct action.

Resolves: #382