[BUGFIX] Customer always gets correct shipping
If shipping costs differs between two countries
the customer could get wrong shipping costs by
choosing another country in the shipping address
and afterwards checking the checkbox "shipping
is same as billing".

This is fixed by using the billing country for the
shipping costs when the customer chooses "shipping
is same as billing".

Furthermore the checkbox is always correct checked
by augmenting the condition. Before the checkbox was
not check on first load but the shipping address
fields were not shown.

Solves #363