[TASK] Reimplement `TitleTagViewHelper`
Adds a ViewHelper to set the page title which
can be used by other extensions for EXT:cart.

The implementation is copied from
EXT:sf_event_mgt. Compared to an implementation
without a ViewHelper (the title could also be set
within the controller) this approach enables other
integrators to easily adapt the page title by simply
adding another value to the argument `pageTitle`.

This is a preparation for https://github.com/extcode/cart_products/issues/132

Fixes #480