[TASK] Clean up language files (#13)
* [TASK] Clean up entries of locallang_db.xlf

`locallang_db.xlf` contains only labels for TCA
and FlexForms.

Superfluous entries are deleted.

Harmonize spelling of email (always without
hyphen).

* [TASK] Clean up `locallang.xlf`

Delete not used entries

* [TASK] Separate language keys of BE templates

The backend list view uses its own
entries from `locallang.xlf` which are namespaced.

EXCEPTION: The entries for status of shipping and
payment are still shared. But it's unlikely that
developer might want to change them.

* [TASK] Show correct message when wrong page

Adapt the message when the module is used on a
page without order items. Furhtermore show a
correct message when the filtered search results
in zero results.

Solves #471

* [TASK] Add labels for multi-step checkout

Add missing labels for the buttons which are used
to move between the steps in a multi-step
checkout setup.