[DOCS] Improve docs (#470)
* [DOCS] Improve docs for email

Change structure to make it easier to find
email related settings.

* [DOCS] Improve docs for plugins

* Add example how to include mini cart plugin.
* Add note for order plugin.

* [DOCS] Improve docs for negative taxClassId's

* [DOCS] Improve email section

* Split email addresses from attachments
* Improve content of pages.

* [DOCS] Improve flexible prices section

Add explanations for flexible price
implementations.

* [DOCS] Add undocumented `extra` type `each`

The parameter `extra` of a Service can be set to
`each`. This was not documented before.

* [DOCS] Document tax classes depending on countries

Tax classes can be defined based on countries
which was not documented before.

* [DOCS] Add section about use cases

* [DOCS] Remove section about variants

This section is not really part of EXT:cart
and will be moved to EXT:cart_products.