[BUGFIX] Persist only existing shippingAdress
In multistep checkout the property
`shippingSameAsBilling` might get between the
steps as it does not get persisted by the
`SessionHandler`. As a result the value can become
false although it was set to true by the customer.
This results in orders where an empty shipping
address is persisted.

The value is also stored in the cart object which
is persisted by the `SessionHandler`. This means
we can trust this value instead.

Before persisting the `orderItem` we now transfer
the value from the `cart` and delete the
shippingAddress object if the value
`shippingSameAsBilling` is true.

Fixes: #492