[BUGFIX] Use PRG pattern for multistep checkout
Avoid errors when navigating in the multistep
checkout between different steps with the browser
navigation (back and forward buttons of the
browser) by redirecting to the same page.

To avoid that the customer can go back to the
last step of the multistep checkout after sending
the order we also redirect to the first page of
the checkout if there are no products in the cart.

Fixes: #518