[TASK] Add labels for gender
Provide labels for gender which makes it
easier for integrators to set up
radio buttons for the salutation instead
of using the default text input field.

With these labels in place a tutorial can
be added to the documentation how
integrators can replace the default input
field with radio buttons. Furthermore another
step could be to add an opt-in solution to
switch to radio button for the addresses.