[FEATURE]  Make Shipping methods be available based on cart conditions (#545)
Implemented the ability to configure conditions in typoscript setup to switch shipping methods
based on products in the cart. This allows e.g. for dynamic application of shipping methods,
ensuring that shipping method A (free) is used by default, and shipping method B (with actual
costs) is applied when physical products are detected in the cart.

Fixes: #545