[BUGFIX] Use ContextualFeedbackServerity in addFlashMessage function
In order to better understand the process in the controller, some code blocks
have been moved to new private methods.
Furthermore, the ENUM ContextualFeedbackSeverity was used directly instead of
the serialised array. A method marked as internal is used, but it is not
expected that this will change within a major version. However, this makes it
much easier to read in the controller.

Related: #646