[BUGFIX] Adapt widget configuration to interface
The change https://review.typo3.org/c/Packages/TYPO3.CMS/+/89562
removed the property `additionalCssClasses` from
the `WidgetConfiguration` which results in an
error when setting this as argument in the
configuration.

This argument was according to the commit message
never used which means removing it does not
result in any changes.

Resolves: #673